refactor(react): tweak focus styles on select fields
