+++ packages/react/src/theme/componentStyles/form.ts
@@ -2,6 +2,7 @@
 import { transparentize } from 'polished';
 import type { DefaultTheme } from 'styled-components';
 import type { ComponentThemeConfig } from '../../system';
+import formSelect from './form/select';
 import formToggle from './form/toggle';
 
 // TODO: move to `componentStyles/form/utils.ts`
@@ -20,19 +21,6 @@ const inputBase = (theme: DefaultTheme, rounded = false) => ({
   '--aks-shadow': '0 0 #0000',
 });
 
-const inputSelectBase = (theme: DefaultTheme) => ({
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'space-between',
-  width: '100%',
-  paddingLeft: '16px',
-  backgroundColor: theme.colors.grey01,
-  border: '1px solid',
-  borderColor: theme.colors.grey04,
-  borderRadius: '8px',
-  outline: 'none',
-});
-
 // TODO: move to `componentStyles/form/utils.ts`
 const inputFocusBase = (theme: DefaultTheme) => ({
   outline: '2px solid transparent',
@@ -41,7 +29,7 @@ const inputFocusBase = (theme: DefaultTheme) => ({
   '--aks-ring-width': '2px',
   '--aks-ring-offset-width': '0px',
   '--aks-ring-offset-color': '#fff',
-  '--aks-ring-color': transparentize(0.7, theme.colors.blue07),
+  '--aks-ring-color': transparentize(0.7, theme.colors.blue06),
   '--aks-ring-offset-shadow': 'var(--aks-ring-inset) 0 0 0 var(--aks-ring-offset-width) var(--aks-ring-offset-color)',
   '--aks-ring-shadow': `var(--aks-ring-inset) 0 0 0 calc(var(--aks-ring-width) + var(--aks-ring-offset-width)) var(--aks-ring-color)`,
   'box-shadow': 'var(--aks-ring-offset-shadow), var(--aks-ring-shadow), var(--aks-shadow)',
@@ -208,8 +196,8 @@ const inputTags: ComponentThemeConfig = {
           },
           '&:focus-within, &:focus, &:active': {
             '--aks-form-bg-color': theme.colors.greylight01,
-            '--aks-form-border-color': theme.colors.blue07,
-            '--aks-ring-color': transparentize(0.7, theme.colors.blue07),
+            '--aks-form-border-color': theme.colors.blue06,
+            '--aks-ring-color': transparentize(0.7, theme.colors.blue06),
           },
           '&::placeholder': {
             color: theme.colors.greymed01,
@@ -219,12 +207,12 @@ const inputTags: ComponentThemeConfig = {
       focused: ({ theme }: { theme: DefaultTheme }) => ({
         '&:not([disabled])': {
           '--aks-form-bg-color': theme.colors.greylight01,
-          '--aks-form-border-color': theme.colors.blue07,
-          '--aks-ring-color': transparentize(0.7, theme.colors.blue07),
+          '--aks-form-border-color': theme.colors.blue06,
+          '--aks-ring-color': transparentize(0.7, theme.colors.blue06),
           '&:hover, &:focus-within, &:focus, &:active': {
             '--aks-form-bg-color': theme.colors.greylight01,
-            '--aks-form-border-color': theme.colors.blue07,
-            '--aks-ring-color': transparentize(0.7, theme.colors.blue07),
+            '--aks-form-border-color': theme.colors.blue06,
+            '--aks-ring-color': transparentize(0.7, theme.colors.blue06),
           },
           '&::placeholder': {
             color: theme.colors.greymed01,
@@ -262,54 +250,6 @@ const inputTags: ComponentThemeConfig = {
   },
 };
 
-const inputSelect: ComponentThemeConfig = {
-  baseStyle: ({ theme }: { theme: DefaultTheme }) => ({
-    ...inputSelectBase(theme),
-    '&[disabled]': {
-      background: theme.colors.greylight02,
-      border: '1px solid',
-      borderColor: transparentize(0.5, theme.colors.greylight05),
-      color: theme.colors.greymed01,
-      cursor: 'not-allowed',
-    },
-  }),
-  propToScaleMap: [
-    ['size', 'sizes'],
-    ['variant', 'variants'],
-  ],
-  scales: {
-    sizes: {
-      md: {
-        height: '32px',
-      },
-      lg: {
-        height: '40px',
-      },
-    },
-    variants: {
-      default: ({ theme }: { theme: DefaultTheme }) => ({
-        '&:not([disabled])': {
-          '&:focus, &:active': {
-            outline: 'none',
-            borderColor: theme.colors.blue06,
-            boxShadow: `0 0 0 2px ${transparentize(0.7, theme.colors.blue03)}`,
-          },
-          '&:hover': {
-            borderColor: theme.colors.greymed01,
-            background: theme.colors.greylight03,
-          },
-        },
-      }),
-      error: ({ theme }: { theme: DefaultTheme }) => ({
-        '&:not([disabled])': {
-          background: theme.colors.red01,
-          borderColor: theme.colors.red07,
-        },
-      }),
-    },
-  },
-};
-
 // TODO: move to `componentStyles/form/message.ts`
 const inputMessage: ComponentThemeConfig = {
   propToScaleMap: [['variant', 'variants']],
@@ -339,7 +279,7 @@ const form = {
   inputTextarea,
   inputTags,
   inputMessage,
-  inputSelect,
+  ...formSelect,
   ...formToggle,
 };
